app/check_models.py
# check_models.py
import os
import google.generativeai as genai
from dotenv import load_dotenv

load_dotenv()

api_key = os.getenv("GOOGLE_API_KEY")
if not api_key:
    print("Kein API Key gefunden!")
    exit()

genai.configure(api_key=api_key)

print("Verfügbare Modelle für generateContent:")
for m in genai.list_models():
    if 'generateContent' in m.supported_generation_methods:
        print(f"- {m.name}")